+++ stringtools.h
@@ -108,7 +108,7 @@ std::string UnescapeHTML(const std::string &html);
 std::wstring UnescapeHTML(const std::wstring &html);
 std::string PrettyPrintBytes(_i64 bytes);
 std::string PrettyPrintSpeed(size_t bps);
-std::string PrettyPrintTime(unsigned int ms);
+std::string PrettyPrintTime(int64 ms);
 std::string EscapeParamString(const std::string &pStr);
 
 namespace
